refactor(project): restructure packages
