+++ agents/drafting_agent.py
@@ -3,8 +3,8 @@ import logging
 from typing import Any, Dict, List, Optional, Tuple
 
 import config
+from core.llm_interface import count_tokens, llm_service, truncate_text_by_tokens
 from kg_maintainer.models import SceneDetail
-from llm_interface import count_tokens, llm_service, truncate_text_by_tokens
 from prompt_renderer import render_prompt
 
 logger = logging.getLogger(__name__)
