+++ agents/finalize_agent.py
@@ -5,11 +5,11 @@ from typing import Any, Dict, Optional, TypedDict
 
 import numpy as np
 import structlog
+from agents.kg_maintainer_agent import KGMaintainerAgent
 
+from core.llm_interface import llm_service
 from data_access import chapter_queries, kg_queries
 from kg_maintainer.models import CharacterProfile, WorldItem
-from kg_maintainer_agent import KGMaintainerAgent
-from llm_interface import llm_service
 from parsing_utils import parse_rdf_triples_with_rdflib
 
 logger = structlog.get_logger(__name__)
