+++ agents/patch_validation_agent.py
@@ -2,8 +2,8 @@ import logging
 from typing import Dict, List, Optional, Tuple
 
 import config
+from core.llm_interface import llm_service
 from kg_maintainer.models import PatchInstruction, ProblemDetail
-from llm_interface import llm_service
 from prompt_renderer import render_prompt
 
 logger = logging.getLogger(__name__)
