+++ agents/planner_agent.py
@@ -5,15 +5,15 @@ import re
 from typing import Any, Dict, List, Optional, Tuple
 
 import config
+from core.llm_interface import llm_service
 from data_access import chapter_queries
-from llm_interface import llm_service
-from prompt_renderer import render_prompt
+from kg_maintainer.models import CharacterProfile, SceneDetail, WorldItem
 from prompt_data_getters import (
     get_character_state_snippet_for_prompt,
     get_reliable_kg_facts_for_drafting_prompt,
     get_world_state_snippet_for_prompt,
 )
-from kg_maintainer.models import CharacterProfile, SceneDetail, WorldItem
+from prompt_renderer import render_prompt
 
 logger = logging.getLogger(__name__)
