+++ data_access/__init__.py
@@ -13,8 +13,8 @@ from .chapter_queries import (
 from .character_queries import (
     get_character_info_for_snippet_from_db,
     get_character_profile_by_name,
-    resolve_character_name,
     get_character_profiles_from_db,
+    resolve_character_name,
     sync_characters,
 )
 from .character_queries import (
@@ -31,9 +31,9 @@ from .plot_queries import get_plot_outline_from_db, save_plot_outline_to_db
 from .world_queries import (
     get_world_building_from_db,
     get_world_elements_for_snippet_from_db,
+    get_world_item_by_id,
     get_world_item_by_name,
     resolve_world_name,
-    get_world_item_by_id,
     sync_world_items,
 )
 from .world_queries import (
