+++ data_access/cypher_builders/__init__.py
@@ -1,4 +1,4 @@
-from .character_cypher import generate_character_node_cypher, TRAIT_NAME_TO_CANONICAL
+from .character_cypher import TRAIT_NAME_TO_CANONICAL, generate_character_node_cypher
 from .world_cypher import generate_world_element_node_cypher
 
 __all__ = [
