+++ data_access/cypher_builders/character_cypher.py
@@ -1,4 +1,5 @@
 from typing import Any, Dict, List, Tuple
+
 import structlog
 
 import config
