+++ initialization/bootstrappers/character_bootstrapper.py
@@ -1,10 +1,12 @@
 import asyncio
 from typing import Any, Coroutine, Dict, Optional, Tuple
+
 import structlog
 
 import config
 import utils
 from kg_maintainer.models import CharacterProfile
+
 from .common import bootstrap_field
 
 logger = structlog.get_logger(__name__)
