+++ initialization/bootstrappers/plot_bootstrapper.py
@@ -1,9 +1,11 @@
 import asyncio
 from typing import Any, Coroutine, Dict, Optional, Tuple
+
 import structlog
 
 import config
 import utils
+
 from .common import bootstrap_field
 
 logger = structlog.get_logger(__name__)
