+++ kg_maintainer/__init__.py
@@ -1,5 +1,9 @@
 """Package consolidating KG maintainer utilities."""
 
+from .merge import (
+    merge_character_profile_updates,
+    merge_world_item_updates,
+)
 from .models import (
     AgentStateData,
     CharacterProfile,
@@ -12,10 +16,6 @@ from .models import (
 from .parsing import (
     parse_unified_character_updates,
     parse_unified_world_updates,
-)
-from .merge import (
-    merge_character_profile_updates,
-    merge_world_item_updates,
 )
 
 __all__ = [
