+++ main.py
@@ -1,8 +1,8 @@
 import asyncio
-
-from orchestration.nana_orchestrator import NANA_Orchestrator, setup_logging_nana
-from core_db.base_db_manager import neo4j_manager
 import logging
+
+from core.db_manager import neo4j_manager
+from orchestration.nana_orchestrator import NANA_Orchestrator, setup_logging_nana
 
 logger = logging.getLogger(__name__)
