+++ models/__init__.py
@@ -1,19 +1,19 @@
 """Central package for SAGA data models."""
 
 from .agent_models import (
-    SceneDetail,
-    ProblemDetail,
     EvaluationResult,
     PatchInstruction,
+    ProblemDetail,
+    SceneDetail,
 )
 from .kg_models import CharacterProfile, WorldItem
 from .user_input_models import (
-    NovelConceptModel,
-    RelationshipModel,
-    ProtagonistModel,
     KeyLocationModel,
-    SettingModel,
+    NovelConceptModel,
     PlotElementsModel,
+    ProtagonistModel,
+    RelationshipModel,
+    SettingModel,
     UserStoryInputModel,
     user_story_to_objects,
 )
