+++ processing/context_generator.py
@@ -9,17 +9,16 @@ import logging
 from typing import Any, List, Optional
 
 import config
-
-from data_access import (
-    chapter_queries,
-)  # For chapter data and similarity search
-from llm_interface import (
+from core.llm_interface import (
     count_tokens,
     llm_service,
     truncate_text_by_tokens,
 )  # MODIFIED
-from prompt_data_getters import get_reliable_kg_facts_for_drafting_prompt
+from data_access import (
+    chapter_queries,
+)  # For chapter data and similarity search
 from kg_maintainer.models import SceneDetail
+from prompt_data_getters import get_reliable_kg_facts_for_drafting_prompt
 
 logger = logging.getLogger(__name__)
