+++ processing/revision_logic.py
@@ -11,6 +11,8 @@ from typing import Any, Dict, List, Optional, Tuple
 
 import config
 import utils  # For numpy_cosine_similarity, find_semantically_closest_segment, AND find_quote_and_sentence_offsets_with_spacy, format_scene_plan_for_prompt
+from agents.patch_validation_agent import PatchValidationAgent
+from core.llm_interface import count_tokens, llm_service, truncate_text_by_tokens
 from kg_maintainer.models import (
     CharacterProfile,
     EvaluationResult,
@@ -19,8 +21,6 @@ from kg_maintainer.models import (
     SceneDetail,
     WorldItem,
 )
-from llm_interface import count_tokens, llm_service, truncate_text_by_tokens
-from patch_validation_agent import PatchValidationAgent
 
 logger = logging.getLogger(__name__)
 utils.load_spacy_model_if_needed()  # Ensure spaCy model is loaded when this module is imported
