--- story_models.py
@@ -1,3 +1,0 @@
-"""Deprecated module kept for backward compatibility."""
-
-from models.user_input_models import *  # noqa: F401,F403
