+++ tests/test_kg_heal.py
@@ -1,5 +1,6 @@
-import pytest
 from unittest.mock import AsyncMock
+
+import pytest
 
 from data_access import kg_queries
