+++ yaml_parser.py
@@ -1,7 +1,8 @@
 # yaml_parser.py
-import yaml
 import logging
-from typing import Dict, Any, Optional
+from typing import Any, Dict, Optional
+
+import yaml
 
 logger = logging.getLogger(__name__)
